Move classes out of `reflect` package
