+++ server/src/main/java/io/spine/server/command/CommandHandler.java
@@ -30,9 +30,8 @@ import io.spine.core.Event;
 import io.spine.protobuf.AnyPacker;
 import io.spine.server.commandbus.CommandDispatcher;
 import io.spine.server.event.EventBus;
+import io.spine.server.model.HandlerMethod;
 import io.spine.server.model.Model;
-import io.spine.server.reflect.CommandHandlerMethod;
-import io.spine.server.reflect.HandlerMethod;
 import io.spine.string.Stringifiers;
 import io.spine.type.MessageClass;
 import io.spine.util.Logging;
