+++ server/src/main/java/io/spine/server/command/CommandHandlerClass.java
@@ -24,7 +24,6 @@ import io.spine.annotation.Internal;
 import io.spine.core.CommandClass;
 import io.spine.server.model.HandlerClass;
 import io.spine.server.model.MessageHandlerMap;
-import io.spine.server.reflect.CommandHandlerMethod;
 
 import java.util.Set;
