+++ server/src/main/java/io/spine/server/event/EventSubscriberClass.java
@@ -24,7 +24,6 @@ import io.spine.annotation.Internal;
 import io.spine.core.EventClass;
 import io.spine.server.model.HandlerClass;
 import io.spine.server.model.MessageHandlerMap;
-import io.spine.server.reflect.EventSubscriberMethod;
 
 import java.util.Set;
