+++ server/src/main/java/io/spine/server/model/HandlerMethodFailedException.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.reflect;
+package io.spine.server.model;
 
 import com.google.protobuf.GeneratedMessageV3;
 import com.google.protobuf.Message;
