+++ server/src/main/java/io/spine/server/procman/ProcessManager.java
@@ -28,13 +28,13 @@ import io.spine.core.Event;
 import io.spine.core.EventEnvelope;
 import io.spine.core.MessageEnvelope;
 import io.spine.core.RejectionEnvelope;
+import io.spine.server.command.CommandHandlerMethod;
 import io.spine.server.command.CommandHandlingEntity;
 import io.spine.server.commandbus.CommandBus;
+import io.spine.server.event.EventReactorMethod;
+import io.spine.server.model.HandlerMethod;
 import io.spine.server.model.Model;
-import io.spine.server.reflect.CommandHandlerMethod;
-import io.spine.server.reflect.EventReactorMethod;
-import io.spine.server.reflect.HandlerMethod;
-import io.spine.server.reflect.RejectionReactorMethod;
+import io.spine.server.rejection.RejectionReactorMethod;
 import io.spine.validate.ValidatingBuilder;
 
 import java.util.List;
