+++ server/src/main/java/io/spine/server/projection/ProjectionClass.java
@@ -22,9 +22,9 @@ package io.spine.server.projection;
 
 import io.spine.annotation.Internal;
 import io.spine.core.EventClass;
+import io.spine.server.event.EventSubscriberMethod;
 import io.spine.server.model.EntityClass;
 import io.spine.server.model.MessageHandlerMap;
-import io.spine.server.reflect.EventSubscriberMethod;
 
 import java.util.Set;
