+++ server/src/test/java/io/spine/server/aggregate/EventApplierMethodShould.java
@@ -18,14 +18,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.reflect;
+package io.spine.server.aggregate;
 
 import com.google.common.testing.NullPointerTester;
 import com.google.protobuf.Any;
 import com.google.protobuf.StringValue;
 import io.spine.core.CommandContext;
-import io.spine.server.aggregate.Aggregate;
-import io.spine.server.aggregate.Apply;
+import io.spine.server.model.HandlerMethod;
 import io.spine.test.reflect.event.RefProjectCreated;
 import io.spine.testdata.Sample;
 import io.spine.validate.StringValueVBuilder;
