+++ server/src/test/java/io/spine/server/commandbus/CommandHandlerShould.java
@@ -31,8 +31,8 @@ import io.spine.core.EventEnvelope;
 import io.spine.core.Events;
 import io.spine.server.BoundedContext;
 import io.spine.server.event.EventBus;
-import io.spine.server.reflect.given.CommandHandlerTestEnv.EventCatcher;
-import io.spine.server.reflect.given.CommandHandlerTestEnv.TestCommandHandler;
+import io.spine.server.event.given.CommandHandlerTestEnv.EventCatcher;
+import io.spine.server.event.given.CommandHandlerTestEnv.TestCommandHandler;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
