+++ server/src/test/java/io/spine/server/model/HandlerMethodShould.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.server.reflect;
+package io.spine.server.model;
 
 import com.google.protobuf.BoolValue;
 import com.google.protobuf.Empty;
