+++ server/src/test/java/io/spine/server/model/MessageHandlerMapShould.java
@@ -22,15 +22,14 @@ package io.spine.server.model;
 
 import io.spine.core.CommandClass;
 import io.spine.server.command.Assign;
-import io.spine.server.reflect.CommandHandlerMethod;
-import io.spine.server.reflect.DuplicateHandlerMethodException;
+import io.spine.server.command.CommandHandlerMethod;
 import io.spine.test.event.ProjectCreated;
 import io.spine.test.event.command.CreateProject;
 import org.junit.Test;
 
 public class MessageHandlerMapShould {
 
-    @Test(expected = DuplicateHandlerMethodException.class)
+    @Test(expected = DuplicateHandlerMethodError.class)
     public void not_allow_duplicating_message_classes() {
         new MessageHandlerMap<CommandClass, CommandHandlerMethod>
                 (HandlerWithDuplicatingMethods.class, CommandHandlerMethod.factory());
